--- process/transactions/datafield/parseRelayed.go
@@ -1,71 +1,0 @@
-package datafield
-
-import (
-	"encoding/json"
-	"github.com/ElrondNetwork/elrond-go-core/core"
-	"github.com/ElrondNetwork/elrond-go-core/data/transaction"
-)
-
-const (
-	numArgsRelayedV2 = 4
-)
-
-func (odp *operationDataFieldParser) ParseRelayed(function string, args [][]byte, receiver []byte) *ResponseParseData {
-	if len(args) == 0 {
-		return &ResponseParseData{
-			IsRelayed: true,
-		}
-	}
-
-	tx, ok := extractInnerTx(function, args, receiver)
-	if !ok {
-		return &ResponseParseData{
-			IsRelayed: true,
-		}
-	}
-
-	res := odp.parse(tx.Data, tx.SndAddr, tx.RcvAddr, true)
-	if res.IsRelayed {
-		return &ResponseParseData{
-			IsRelayed: true,
-		}
-	}
-
-	receivers := []string{odp.pubKeyConverter.Encode(tx.RcvAddr)}
-	receiversShardID := []uint32{odp.shardCoordinator.ComputeId(tx.RcvAddr)}
-	if res.Operation == core.BuiltInFunctionMultiESDTNFTTransfer || res.Operation == core.BuiltInFunctionESDTNFTTransfer {
-		receivers = res.Receivers
-		receiversShardID = res.ReceiversShardID
-	}
-
-	return &ResponseParseData{
-		Operation:        res.Operation,
-		Function:         res.Function,
-		ESDTValues:       res.ESDTValues,
-		Tokens:           res.Tokens,
-		Receivers:        receivers,
-		ReceiversShardID: receiversShardID,
-		IsRelayed:        true,
-	}
-}
-
-func extractInnerTx(function string, args [][]byte, receiver []byte) (*transaction.Transaction, bool) {
-	tx := &transaction.Transaction{}
-
-	if function == core.RelayedTransaction {
-		err := json.Unmarshal(args[0], &tx)
-
-		return tx, err == nil
-	}
-
-	if len(args) != numArgsRelayedV2 {
-		return nil, false
-	}
-
-	// sender of the inner tx is the receiver of the relayed tx
-	tx.SndAddr = receiver
-	tx.RcvAddr = args[0]
-	tx.Data = args[2]
-
-	return tx, true
-}
