+++ process/transactions/datafield/parser.go
@@ -1,19 +1,23 @@
 package datafield
 
 import (
+	"encoding/json"
+	"math/big"
+
 	indexer "github.com/ElrondNetwork/elastic-indexer-go"
 	"github.com/ElrondNetwork/elastic-indexer-go/converters"
 	"github.com/ElrondNetwork/elrond-go-core/core"
 	"github.com/ElrondNetwork/elrond-go-core/core/check"
+	"github.com/ElrondNetwork/elrond-go-core/data/transaction"
 	vmcommon "github.com/ElrondNetwork/elrond-vm-common"
 	"github.com/ElrondNetwork/elrond-vm-common/parsers"
-	"math/big"
 )
 
 const (
 	operationTransfer                 = `transfer`
 	minArgumentsQuantityOperationESDT = 2
 	minArgumentsQuantityOperationNFT  = 3
+	numArgsRelayedV2                  = 4
 )
 
 type operationDataFieldParser struct {
@@ -83,7 +87,7 @@ func (odp *operationDataFieldParser) parse(dataField []byte, sender, receiver []
 				IsRelayed: true,
 			}
 		}
-		return odp.ParseRelayed(function, args, receiver)
+		return odp.parseRelayed(function, args, receiver)
 	}
 
 	if function != "" && core.IsSmartContractAddress(receiver) {
@@ -91,6 +95,66 @@ func (odp *operationDataFieldParser) parse(dataField []byte, sender, receiver []
 	}
 
 	return responseParse
+}
+
+func (odp *operationDataFieldParser) parseRelayed(function string, args [][]byte, receiver []byte) *ResponseParseData {
+	if len(args) == 0 {
+		return &ResponseParseData{
+			IsRelayed: true,
+		}
+	}
+
+	tx, ok := extractInnerTx(function, args, receiver)
+	if !ok {
+		return &ResponseParseData{
+			IsRelayed: true,
+		}
+	}
+
+	res := odp.parse(tx.Data, tx.SndAddr, tx.RcvAddr, true)
+	if res.IsRelayed {
+		return &ResponseParseData{
+			IsRelayed: true,
+		}
+	}
+
+	receivers := []string{odp.pubKeyConverter.Encode(tx.RcvAddr)}
+	receiversShardID := []uint32{odp.shardCoordinator.ComputeId(tx.RcvAddr)}
+	if res.Operation == core.BuiltInFunctionMultiESDTNFTTransfer || res.Operation == core.BuiltInFunctionESDTNFTTransfer {
+		receivers = res.Receivers
+		receiversShardID = res.ReceiversShardID
+	}
+
+	return &ResponseParseData{
+		Operation:        res.Operation,
+		Function:         res.Function,
+		ESDTValues:       res.ESDTValues,
+		Tokens:           res.Tokens,
+		Receivers:        receivers,
+		ReceiversShardID: receiversShardID,
+		IsRelayed:        true,
+	}
+}
+
+func extractInnerTx(function string, args [][]byte, receiver []byte) (*transaction.Transaction, bool) {
+	tx := &transaction.Transaction{}
+
+	if function == core.RelayedTransaction {
+		err := json.Unmarshal(args[0], &tx)
+
+		return tx, err == nil
+	}
+
+	if len(args) != numArgsRelayedV2 {
+		return nil, false
+	}
+
+	// sender of the inner tx is the receiver of the relayed tx
+	tx.SndAddr = receiver
+	tx.RcvAddr = args[0]
+	tx.Data = args[2]
+
+	return tx, true
 }
 
 func parseBlockingOperationESDT(args [][]byte, funcName string) *ResponseParseData {
